Split trigger conformance tests
